refactor: 整理并优化大量游戏逻辑代码
本次提交包含多项核心优化与重构：
1. 商队系统：替换随机数生成逻辑，使用基于帧和交易双方的哈希值生成更可控的暴击概率
2. 修复类型安全：将循环变量从int改为size_t，消除类型转换警告
3. 资源组件：添加灵石上限常量并在添加资源时自动截断
4. NPC文档：新增社交性格维度说明
5. 谣言与社交组件：为结构体添加默认构造值，移除调试日志
6. 行为树重构：
   - 移除冗余的可执行检测函数，简化执行表配置
   - 新增聊天行为逻辑，完善社交互动流程
   - 重构经济策略逻辑，优化物资储备与贸易禁运算法
7. 市场注册表：修复经济缓存逻辑，支持按宗门区分缓存数据
8. 清理废弃代码：移除旧版BehaviorTag相关代码，简化决策日志宏

+++ src/server/game/ecs/components/CommandDelegationComponent.h
@@ -125,7 +125,7 @@ struct CommandDelegationComponent : public ECS::ComponentBase<CommandDelegationC
     void clearSlot(uint32_t parentCommandId) {
         for (size_t i = 0; i < slotCount; ++i) {
             if (slots[i].parentCommandId == parentCommandId) {
-                for (size_t j = i; j < slotCount - 1; ++j) {
+                for (size_t j = i; j < static_cast<size_t>(slotCount) - 1; ++j) {
                     slots[j] = slots[j + 1];
                 }
                 slots[slotCount - 1] = {};
